+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - unhandled promise rejection on initial connect timeout
 - missing documentation for VrpcNative
 - bug on new `vrpcOn` feature
+- bug while unregister event listeners via "off" or "removeAllListeners"
 
 ### Changed